Create file directory structure via Vite and React
Used the CLI to set up a new React project with
the files provided by the Vite template. The
structure is based on the one we used in the CV
app for the Odin Project. This means we added a
components folder inside src and moved all the
React component files there. We also created a
styles folder inside src and moved all the CSS
files there. Finally, we updated the import
statements in the JS files to reflect the new
paths.