Modify Scoreboard and Shuffle components
Added Deck component to load under the Scoreboard component inspired
from the Amphibia Memory Game layout. Refactored the Shuffle component
by renaming it to the Deck component as that is a better name for what
the component does currently. The Deck component now fetches the starter
Pokemon from PokeAPI and puts them into their own array, displays the
cards representing them and can shuffle those "cards".

Wrote some rough draft code representing the Card component itself which
fetches both the image and the text for a (different) starter Pokemon
every time the component is called.

Next steps:

Added TODO dev comments for refactoring conditional statement in the
Scoreboard component. Added more dev comments concerning the shuffle
code in the Deck component and whether or not it even needs to be an
Effect (it isn't in the Amphibia game and the React docs seem to say the
same thing: that it doesn't need to be an Effect).